fix: Move handler and exceptions in the controller package
