+++ src/main/java/com/java/test/junior/controller/FileNotFoundException.java
@@ -1,4 +1,4 @@
-package com.java.test.junior.exception;
+package com.java.test.junior.controller;
 
 public class FileNotFoundException extends RuntimeException {
     public FileNotFoundException(String message) {
