+++ src/main/java/com/java/test/junior/controller/IllegalArgumentException.java
@@ -1,4 +1,4 @@
-package com.java.test.junior.exception;
+package com.java.test.junior.controller;
 
 public class IllegalArgumentException extends RuntimeException{
     public IllegalArgumentException(String message) {
