+++ src/main/java/com/java/test/junior/controller/ProductNotFoundException.java
@@ -1,4 +1,4 @@
-package com.java.test.junior.exception;
+package com.java.test.junior.controller;
 
 public class ProductNotFoundException extends RuntimeException{
     public ProductNotFoundException(String message) {
